bot - 251122.1/requirements.txt
discord.py>=2.3.2
SQLAlchemy>=2.0.0
mysql-connector-python>=8.0.33
aiohttp>=3.9
